Fix step failure message
When a step failed, the step's Kind was being passed in as the format param. This wasn't a valid format and so an Printf error was being printed out resulting in a garbled error. Added a proper format param in front of the Kind. Also extracted the function that runs each step so it could be unit tested properly